cleaned code with unnessicairy variables. Added break function
