+++ code/src/qik_2s12v10.cpp
@@ -13,47 +13,39 @@ namespace r2d2::moving_platform {
 
     qik_2s12v10_c::qik_2s12v10_c(r2d2::uart_ports_c uart_port,
                                  unsigned int baudRate,
-                                 hwlib::pin_out *_reset_pin)
-        : reset_pin(_reset_pin), usart_bus(baudRate, uart_port) {
-    }
+                                 hwlib::pin_out *_reset_pin): 
+        reset_pin(_reset_pin), 
+        usart_bus(baudRate, uart_port) 
+        {}
 
     void qik_2s12v10_c::set_speed(const int8_t &_speed) {
         // Todo: depending on the motor configuration, 127 may not be the
         // maximum value
         // Todo: change the speed variable to the actual speed.
-        uint8_t speed_byte = 0;
         if (_speed >= 0) {
-            speed_byte = _speed;
-            usart_bus << qik_motor_m0_set_forward << speed_byte;
-            usart_bus << qik_motor_m1_set_forward << speed_byte;
+            usart_bus << qik_motor_m0_set_forward << _speed;
+            usart_bus << qik_motor_m1_set_forward << _speed;
         } else {
-            speed_byte = -1 * _speed;
-            usart_bus << qik_motor_m0_set_reverse << speed_byte;
-            usart_bus << qik_motor_m1_set_reverse << speed_byte;
+            usart_bus << qik_motor_m0_set_reverse << (-1 * _speed);
+            usart_bus << qik_motor_m1_set_reverse << (-1 * _speed);
         }
     }
     void qik_2s12v10_c::set_m0_speed(const int8_t &_speed) {
         // Todo: depending on the motor configuration, 127 may not be the
         // maximum value
-        uint8_t speed_byte = 0;
         if (_speed >= 0) {
-            speed_byte = _speed;
-            usart_bus << qik_motor_m0_set_forward << speed_byte;
+            usart_bus << qik_motor_m0_set_forward << _speed;
         } else {
-            speed_byte = -1 * _speed;
-            usart_bus << qik_motor_m0_set_reverse << speed_byte;
+            usart_bus << qik_motor_m0_set_reverse << (-1 * _speed);
         }
     }
     void qik_2s12v10_c::set_m1_speed(const int8_t &_speed) {
         // Todo: depending on the motor configuration, 127 may not be the
         // maximum value
-        uint8_t speed_byte = 0;
         if (_speed >= 0) {
-            speed_byte = _speed;
-            usart_bus << qik_motor_m1_set_forward << speed_byte;
+            usart_bus << qik_motor_m1_set_forward << _speed;
         } else {
-            speed_byte = -1 * _speed;
-            usart_bus << qik_motor_m1_set_reverse << speed_byte;
+            usart_bus << qik_motor_m1_set_reverse << (-1 * _speed);
         }
     }
 
@@ -67,11 +59,9 @@ namespace r2d2::moving_platform {
         // setup the uart communication
         usart_bus << qik_autodetect_baudrate;
     }
-    void qik_2s12v10_c::set_m0_brake(const uint8_t &brake) {
-    }
-    void qik_2s12v10_c::set_m1_brake(const uint8_t &brake) {
-    }
-    void qik_2s12v10_c::set_brakes(const uint8_t &brake) {
+
+    void qik_2s12v10_c::brake(){
+        set_speed(0);
     }
 
     uint8_t qik_2s12v10_c::get_error() {
@@ -80,16 +70,17 @@ namespace r2d2::moving_platform {
         }                           // clear the buffer
         usart_bus << qik_get_error; // send request
         while (!usart_bus.available()) {
+            hwlib::wait_ms(50);     // don't use all time
         }                           // wait for answer
         return usart_bus.receive(); // return answer
     }
-    uint8_t
-    qik_2s12v10_c::get_configuration_parameter(const uint8_t &parameter) {
+    uint8_t qik_2s12v10_c::get_configuration_parameter(const uint8_t &parameter) {
         while (usart_bus.available() > 0) {
             usart_bus.receive();
         }                                                   // clear the buffer
         usart_bus << qik_get_config_parameter << parameter; // send request
         while (!usart_bus.available()) {
+            hwlib::wait_ms(50);     // don't use all time
         }                           // wait for answer
         return usart_bus.receive(); // return answer
     }
